// src/lib/game-data/silk-data/tool-crest-silk.generated.ts
/**
 * This file is auto-generated by a script (See ./scripts/*). Do not edit directly.
 */

export type ToolCrestNameSilk =
	| 'Hunter'
	| 'Hunter_v2'
	| 'Hunter_v3'
	| 'Reaper'
	| 'Wanderer'
	| 'Warrior'
	| 'Witch'
	| 'Toolmaster'
	| 'Spell'
	| 'Cursed'
	| 'Cloakless';

const list = [
	{
		id: 'Hunter',
		displayName: 'Tools.CREST_HUNTER_NAME',
		crestSprite: {
			name: 'Crest__0014_hunter_lvl2_37_427x512_54',
			nameShort: 'Crest__0014_hunter_lvl2',
			size: {
				x: 427,
				y: 512,
			},
			padding: {
				x: 0,
				y: 82.07612,
				z: 0,
				w: 0,
			},
		},
	},
	{
		id: 'Hunter_v2',
		displayName: 'Tools.CREST_HUNTER_NAME',
		crestSprite: {
			name: 'Crest__0014_hunter_lvl3_46_511x649_48',
			nameShort: 'Crest__0014_hunter_lvl3',
			size: {
				x: 511,
				y: 649,
			},
			padding: {
				x: 2.07612038,
				y: 147.076126,
				z: 4.076111,
				w: 2.07611084,
			},
		},
	},
	{
		id: 'Hunter_v3',
		displayName: 'Tools.CREST_HUNTER_NAME',
		crestSprite: {
			name: 'Crest__0014_hunter_lvl4_47_481x663_58',
			nameShort: 'Crest__0014_hunter_lvl4',
			size: {
				x: 481,
				y: 663,
			},
			padding: {
				x: 5.07612038,
				y: 159.076126,
				z: 8.076111,
				w: 9.076111,
			},
		},
	},
	{
		id: 'Reaper',
		displayName: 'Tools.CREST_REAPER_NAME',
		crestSprite: {
			name: 'Crest__0004_reaper_70_579x463_28',
			nameShort: 'Crest__0004_reaper',
			size: {
				x: 579,
				y: 463,
			},
			padding: {
				x: 0,
				y: 0,
				z: 0,
				w: 0,
			},
		},
	},
	{
		id: 'Wanderer',
		displayName: 'Tools.CREST_PILGRIM_NAME',
		crestSprite: {
			name: 'Crest__0006_wanderer_86_433x668_31',
			nameShort: 'Crest__0006_wanderer',
			size: {
				x: 433,
				y: 668,
			},
			padding: {
				x: 0,
				y: 10.07612,
				z: 0,
				w: 54.0267334,
			},
		},
	},
	{
		id: 'Warrior',
		displayName: 'Tools.CREST_WARRIOR_NAME',
		crestSprite: {
			name: 'Crest__0007_warrior_63_627x574_46',
			nameShort: 'Crest__0007_warrior',
			size: {
				x: 627,
				y: 574,
			},
			padding: {
				x: 53.0761223,
				y: 43.0129128,
				z: 53.07611,
				w: 43.0267334,
			},
		},
	},
	{
		id: 'Witch',
		displayName: 'Tools.CREST_WITCH_NAME',
		crestSprite: {
			name: 'Crest__0006_witch_57_580x613_85',
			nameShort: 'Crest__0006_witch',
			size: {
				x: 580,
				y: 613,
			},
			padding: {
				x: 0,
				y: 0,
				z: 0,
				w: 0,
			},
		},
	},
	{
		id: 'Toolmaster',
		displayName: 'Tools.CREST_TOOLMASTER_NAME',
		crestSprite: {
			name: 'Crest__0017_toolmaster_88_560x634_72',
			nameShort: 'Crest__0017_toolmaster',
			size: {
				x: 560,
				y: 634,
			},
			padding: {
				x: 0,
				y: 22.0362854,
				z: 0,
				w: 0,
			},
		},
	},
	{
		id: 'Spell',
		displayName: 'Tools.CREST_SPELL_NAME',
		crestSprite: {
			name: 'Crest__0010_spell_112_597x633_76',
			nameShort: 'Crest__0010_spell',
			size: {
				x: 597,
				y: 633,
			},
			padding: {
				x: 24.0429077,
				y: 5.04290771,
				z: 4.05133057,
				w: 40.07611,
			},
		},
	},
	{
		id: 'Cursed',
		displayName: 'Tools.CREST_CURSE_NAME',
		crestSprite: {
			name: 'cursed_crest_animated0001_40_329x612_65',
			nameShort: 'cursed_crest_animated0001',
			size: {
				x: 329,
				y: 612,
			},
			padding: {
				x: 14.07612,
				y: 22.07612,
				z: 16.07611,
				w: 9.076111,
			},
		},
	},
	{
		id: 'Cloakless',
		displayName: null,
		crestSprite: null,
	},
];
export type ToolCrestSilk = (typeof list)[number];
export const toolCrestSilk = {
	list,
	byName: Object.fromEntries(list.map((it: any) => [it.id, it])) as Record<ToolCrestNameSilk, ToolCrestSilk>,
};

export const toolCrestIdToNameSilk: Map<number, ToolCrestNameSilk> = new Map([
	[1, 'Hunter'],
	[2, 'Hunter_v2'],
	[3, 'Hunter_v3'],
	[4, 'Reaper'],
	[5, 'Wanderer'],
	[6, 'Warrior'],
	[7, 'Witch'],
	[8, 'Toolmaster'],
	[9, 'Spell'],
	[10, 'Cursed'],
	[11, 'Cloakless'],
]);
export const toolCrestNamesSilk: ToolCrestNameSilk[] = toolCrestIdToNameSilk.values().toArray();